// src/schemas/user.schema.ts
import { z } from "zod";

export const signupSchema = z.object({
    name: z.string().min(3,"Name must be at least 3 characters"),
    email: z.email("Invalid email format"),
    password: z.string().min(8,"Password must be at least 8 characters"),
    role: z.enum(['student','instructor']).optional().default('student')
});

export const loginSchema = z.object({
    email: z.email("Invalid email format"),
    password: z.string().min(8,"Password must be at least 8 characters")
});

export type SignupInput = z.infer<typeof signupSchema>;
export type LoginInput = z.infer<typeof loginSchema>;